Refactor SE3 into a proper library, add sample unit test
