+++ src/util/se3/so3.cpp
